backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo[srv]==4.6.1
werkzeug==3.0.4
PyJWT==2.9.0
python-dotenv==1.0.1